Detect request content type
