+++ sdk/core/Azure.Core/src/Pipeline/ResponseClassifier.cs
@@ -3,7 +3,6 @@
 
 using System;
 using System.IO;
-using System.Text;
 
 namespace Azure.Core.Pipeline
 {
@@ -32,44 +31,6 @@ namespace Azure.Core.Pipeline
         {
             var statusKind = response.Status / 100;
             return statusKind == 4 || statusKind == 5;
-        }
-
-        public static bool IsTextResponse(Response response, out Encoding encoding)
-        {
-            const string charsetMarker = "; charset=";
-            const string utf8Charset = "utf-8";
-            const string textContentTypePrefix = "text/";
-            const string jsonSuffix = "json";
-            const string xmlSuffix = "xml";
-
-            var contentType = response.Headers.ContentType;
-            if (contentType == null)
-            {
-                encoding = null;
-                return false;
-            }
-
-            var charsetIndex = contentType.IndexOf(charsetMarker, StringComparison.InvariantCultureIgnoreCase);
-            if (charsetIndex != -1)
-            {
-                ReadOnlySpan<char> charset = contentType.AsSpan().Slice(charsetIndex + charsetMarker.Length);
-                if (charset.StartsWith(utf8Charset.AsSpan(), StringComparison.OrdinalIgnoreCase))
-                {
-                    encoding = Encoding.UTF8;
-                    return true;
-                }
-            }
-
-            if (contentType.StartsWith(textContentTypePrefix) ||
-                contentType.EndsWith(jsonSuffix) ||
-                contentType.EndsWith(xmlSuffix))
-            {
-                encoding = Encoding.UTF8;
-                return true;
-            }
-
-            encoding = null;
-            return false;
         }
     }
 }
