+++ sdk/core/Azure.Core/src/ResponseExceptionExtensions.cs
@@ -45,7 +45,7 @@ namespace Azure.Core
                 .AppendLine(")");
 
             if (response.ContentStream != null &&
-                ResponseClassifier.IsTextResponse(response, out var encoding))
+                ContentTypeUtilities.IsText(response.Headers.ContentType, out var encoding))
             {
                 messageBuilder
                     .AppendLine()
